Move OAuth-endpoints under /oauth2 path instead of /connect
