+++ TeachingRecordSystem/src/TeachingRecordSystem.AuthorizeAccess/Pages/OAuth2/SignOut.cshtml
@@ -1,7 +1,7 @@
-@page "/connect/logout"
+@page "/oauth2/logout"
 @using Microsoft.AspNetCore.Http.Extensions
 @using Microsoft.Extensions.Primitives
-@using TeachingRecordSystem.AuthorizeAccess.Pages.Oidc
+@using TeachingRecordSystem.AuthorizeAccess.Pages.OAuth2
 @addTagHelper *, Joonasw.AspNetCore.SecurityHeaders
 @model SignOutModel
 @{
